+++ packages/pudu/v0.1/src/Pudu/Eval/Operator/Access.hs
@@ -39,7 +39,6 @@ import Pudu.Eval.Value
   , CharMethod (..)
   , Closure (..)
   , MapMethod (..)
-  , RangeMethod (..)
   , SetMethod (..)
   , StringMethod (..)
   , Value (..)